Don't register achievement event on 1.12, fixes deprecation warning
