+++ src/main/java/com/ensifera/animosity/craftirc/CraftIRCListener.java
@@ -7,7 +7,6 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.event.player.AsyncPlayerChatEvent;
-import org.bukkit.event.player.PlayerAchievementAwardedEvent;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerKickEvent;
@@ -169,28 +168,6 @@ final class CraftIRCListener implements Listener {
         msg.setField("realSender", event.getEntity().getName());
         msg.setField("prefix", this.plugin.getPrefix(event.getEntity()));
         msg.setField("suffix", this.plugin.getSuffix(event.getEntity()));
-        msg.post();
-    }
-
-    @EventHandler(priority = EventPriority.MONITOR)
-    public void onPlayerAchievement(PlayerAchievementAwardedEvent event) {
-        if (this.plugin.isHeld(CraftIRC.HoldType.ADVANCEMENTS)) {
-            return;
-        }
-        if (event.getAchievement() == null) {
-            return;
-        }
-        final RelayedMessage msg = this.plugin.newMsg(this.plugin.getEndPoint(this.plugin.cMinecraftTag()), null, "achievement");
-        if (msg == null) {
-            return;
-        }
-        msg.setField("sender", event.getPlayer().getDisplayName());
-        msg.setField("message", event.getAchievement().name());
-        msg.setField("world", event.getPlayer().getWorld().getName());
-        msg.setField("realSender", event.getPlayer().getName());
-        msg.setField("prefix", this.plugin.getPrefix(event.getPlayer()));
-        msg.setField("suffix", this.plugin.getSuffix(event.getPlayer()));
-        msg.doNotColor("message");
         msg.post();
     }
 }
